session-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.0
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.12.1
